feat(core): refactorizar servicios e interfaces de solicitud/cliente y reorganizar tests
se agregan interfaces para cliente y solicitud en la capa de infraestructura
se crean/ajustan servicios para la logica de cliente y cierre de solicitudes
se corrige la estructura de archivos fuente Java para que compile correctamente
se reorganizan los tests en clases separadas por entidad (cliente, solicitud, tecnico)
se limpia ApplicationTests removiendo casos movidos
se actualiza el modelo Cliente con constructor parametrizado
