+++ src/main/java/com/mgcss/infraestrucure/ClienteInterfaz.java
@@ -1,0 +1,5 @@
+package com.mgcss.infraestrucure;
+
+public interface ClienteInterfaz {
+
+}
