--- src/main/java/com/mgcss/infraestrucure/SolicitudInterfaz
@@ -1,14 +1,0 @@
-package com.mgcss.service;
-
-import com.mgcss.domain.Cliente;
-import java.util.List;
-
-
-public interface ServicioCliente {
-    
-    List<Cliente> listarClientes();
-    Cliente save(Cliente nuevoCliente);
-    CLiente consultarCliente(Long id);
-}
-
-
