+++ src/main/java/com/mgcss/infraestrucure/SolicitudInterfaz.java
@@ -1,0 +1,8 @@
+package com.mgcss.infraestrucure;
+
+import com.mgcss.domain.Solicitud;
+
+public interface SolicitudInterfaz {
+    public boolean cerrarSolicitud(Solicitud solicitud) ;
+}
+
