--- src/main/java/com/mgcss/service/ServicioSolicitud
@@ -1,12 +1,0 @@
-public interface Servicio {
-    List<Solicitud> obtenerSolicitudes();
-
-    Solicitud obtenerSolicitudPorId(Long id);
-
-    void registrarSolicitud(Solicitud solicitud);
-    void registrarSolicitud();
-
-    void actualizarSolicitud(Solicitud solicitud);
-
-    void eliminarSolicitud(Long id);  
-}
